fx: move PowerUp
