+++ meson.build
@@ -100,6 +100,7 @@ sources = [
   'src/game/effect_routines/flipmap.c',
   'src/game/effect_routines/flood.c',
   'src/game/effect_routines/lara_effects.c',
+  'src/game/effect_routines/powerup.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/effects/gun.c',
@@ -160,7 +161,6 @@ sources = [
   'src/game/objects/effects/explosion.c',
   'src/game/objects/effects/gunshot.c',
   'src/game/objects/effects/missile.c',
-  'src/game/objects/effects/powerup.c',
   'src/game/objects/effects/raising_block.c',
   'src/game/objects/effects/ricochet.c',
   'src/game/objects/effects/sand.c',
