+++ src/game/objects/ai/centaur.c
@@ -4,9 +4,9 @@
 #include "game/collide.h"
 #include "game/effects/blood.h"
 #include "game/effects/exploding_death.h"
+#include "game/effects/gun.h"
 #include "game/items.h"
 #include "game/lot.h"
-#include "game/effects/gun.h"
 #include "game/people.h"
 #include "game/random.h"
 #include "game/sound.h"
