+++ src/game/objects/ai/natla.c
@@ -4,10 +4,10 @@
 #include "game/box.h"
 #include "game/collide.h"
 #include "game/control.h"
+#include "game/effects/gun.h"
 #include "game/items.h"
 #include "game/lot.h"
 #include "game/music.h"
-#include "game/effects/gun.h"
 #include "game/people.h"
 #include "game/random.h"
 #include "game/sound.h"
