+++ src/global/vars.c
@@ -8,7 +8,7 @@
 #include "game/effect_routines/flipmap.h"
 #include "game/effect_routines/flood.h"
 #include "game/effect_routines/lara_effects.h"
-#include "game/objects/effects/powerup.h"
+#include "game/effect_routines/powerup.h"
 #include "game/objects/effects/raising_block.h"
 #include "game/objects/effects/sand.h"
 #include "game/objects/effects/stairs2slope.h"
@@ -22,7 +22,7 @@ void (*g_EffectRoutines[])(ITEM_INFO *item) = {
     Turn180,          FX_DinoStomp,   FX_LaraNormal,
     FX_Bubbles,       FX_FinishLevel, FX_Earthquake,
     FX_Flood,         RaisingBlock,   Stairs2Slope,
-    DropSand,         PowerUp,        FX_Explosion,
+    DropSand,         FX_PowerUp,     FX_Explosion,
     FX_LaraHandsFree, FX_FlipMap,     FX_LaraDrawRightGun,
     FX_ChainBlock,    FX_Flicker,
 };
